# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My REST API",
        "description": "This is the OpenAPI documentation for my Laravel REST API project.",
        "contact": {
            "name": "Mantas",
            "email": "mantas@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/listings": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Get all approved listings",
                "operationId": "626511cff52fa1165beca4b7480a66c8",
                "responses": {
                    "200": {
                        "description": "List of approved listings"
                    }
                }
            },
            "post": {
                "tags": [
                    "Listings"
                ],
                "summary": "Create a new listing",
                "operationId": "f232854be0ba1c4e83eeda78909de489",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_model_id",
                                    "title",
                                    "description",
                                    "price",
                                    "mileage",
                                    "color"
                                ],
                                "properties": {
                                    "car_model_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "2016 Toyota Corolla"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Well-maintained, low mileage car."
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8500
                                    },
                                    "mileage": {
                                        "type": "integer",
                                        "example": 120000
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Red"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Listing created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/car-models/{modelId}/listings": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Get listings by car model",
                "operationId": "dc61de9f18bf5c713a0d6afdfba56c9e",
                "parameters": [
                    {
                        "name": "modelId",
                        "in": "path",
                        "description": "Car model ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of listings for the given model"
                    },
                    "404": {
                        "description": "Car model not found"
                    }
                }
            }
        },
        "/api/listings/{id}": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Get listing details (with comments)",
                "operationId": "3d6056b16aecf7f8e24443d7e69f9e1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Listing ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listing details with related data"
                    },
                    "404": {
                        "description": "Listing not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Listings"
                ],
                "summary": "Update listing (owner only)",
                "operationId": "f0c3d97b00edab2226ce3a2e80b3c060",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Listing ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated car title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 9500
                                    },
                                    "mileage": {
                                        "type": "integer",
                                        "example": 130000
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Blue"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Listing updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Listing not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Listings"
                ],
                "summary": "Delete listing",
                "operationId": "9f2ef5353184b7ca889c8e4d150cc093",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Listing ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Listing deleted successfully"
                    },
                    "404": {
                        "description": "Listing not found"
                    }
                }
            }
        },
        "/api/listings/{id}/approve": {
            "patch": {
                "tags": [
                    "Listings"
                ],
                "summary": "Approve listing (admin only)",
                "operationId": "db70ce06dfebe5f03cb2bd12dd8474a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Listing ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listing approved successfully"
                    },
                    "404": {
                        "description": "Listing not found"
                    }
                }
            }
        },
        "/api/listings/{id}/reject": {
            "patch": {
                "tags": [
                    "Listings"
                ],
                "summary": "Reject listing (admin only)",
                "operationId": "0696d64ccee64568c8a92c1ba18cc0af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Listing ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "admin_comment": {
                                        "type": "string",
                                        "example": "Listing rejected due to incomplete info"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Listing rejected successfully"
                    },
                    "404": {
                        "description": "Listing not found"
                    }
                }
            }
        }
    },
...
}